refactor(site): dont show select helper when having one proxy
